[LiveChat] - 입장 메시지 중복 현상 수정 완료 - 채팅 보냈을 때 중복 전송 현상 수정 완료

+++ src/main/java/ac/su/suport/livescore/controller/ChatController.java
@@ -34,7 +34,7 @@ public class ChatController {
         message.setRoomId(matchId);
 
         // 채팅방 인원수 세팅
-        long userCount = chatRoomRepository.getUserCount(matchId);
+        int userCount = (int) chatRoomRepository.getUserCount(matchId);
         logger.debug("Current user count in room {}: {}", matchId, userCount);
         message.setUserCount(userCount);
